Fully define images in vars
Before the image specs for packit-service and packit-service-worker were
composed of the image name and the tag, and the later was always
`deployment`.

This made somewhat difficult to deploy the service with *:prod images
locally.

This is made much simpler if we allow providing a full spec for these
images in `vars`.

Signed-off-by: Hunor Csomortáni <csomh@redhat.com>