Separate Redis and Redis Commander deployment
This is to enable deploying Redis and/or Redis Commander on their own,
independent of the other objects in the OpenShift project.

One can use the `TAGS` env var to select tags to be run by `make
deploy`.

For example:

    $ DEPLOYMENT=DEV TAGS="redis,redis-commander" make deploy

So far only "redis" and "redis-commander" are available as tags. `make
tags` can be used to tell all the tags available.

Furthermore: update Redis and Redis Commander related tasks to use
[`apply: yes`] for resource definitions, in order to make these
definitions idempotent. (Previously removing an environment variable
from the DeploymentConfig did not result in a DeploymentConfig change
because of the default merge_type used by k8s.) This requires Ansible
2.9 to run, but it's clearer then tweaking [merge_type].

[`apply: yes`]: https://docs.ansible.com/ansible/latest/modules/k8s_module.html#parameter-apply
[merge_type]: https://docs.ansible.com/ansible/latest/modules/k8s_module.html#parameter-merge_type

Signed-off-by: Hunor Csomortáni <csomh@redhat.com>