nginx: Rework the deployment in order to drop authentication
Rework the nginx deployment, so that the exactly matched /metrics
endpoint is always redirected to pushgateway. Just to be sure, change
the method to be always GET.

All other endpoints should return 404 NOT FOUND.

This way we allow pushing metrics from other pods in the namespace
without authentication to the pushgateway and expose only the /metrics
endpoint externally for unauthenticated scraping.

Signed-off-by: Hunor Csomortáni <csomh@redhat.com>