Control pushing dev images with a dedicated flag
Up until now pushing the dev images depended on the name of the
deployment.

But this task was failing for deployments which were not done on a
cluster started with 'oc cluster up', but with 'minishift', for example.

Instead of relying on the name of the deployment to decide whether to
push the dev images, create a dedicated flag, to allow disabling this
task even when the deployment is done on an environment called 'dev' but
not on a cluster started with 'oc cluster up'.

Signed-off-by: Hunor Csomortáni <csomh@redhat.com>