Improve deleting stale sandbox PVCs
The previous approach checked for stale pods first and deleted the
corresponding PVCs. But as we found out in the meantime, sometimes pods
are successfully removed, but the PVCs requested by them are not. These
PVCs would be never considered for deletion.

Change the approach to first delete stale pods, then identify all the
sandbox PVCs that are mounted in non-stale pods and then delete any
sandbox PVC that is not mounted.

Set up some basic testing.

Signed-off-by: Hunor Csomortáni <csomh@redhat.com>