Increase postgres resources
The prod instance is getting quite hungry
and even the metrics say it's not eating that much,
by giving the pod higher limit the postgres automatically
gets bigger shared_buffers (mem_limit/4) and effective_cache_size
(mem_limit/2) which improves performance.
https://github.com/sclorg/postgresql-container/blob/master/src/root/usr/share/container-scripts/postgresql/common.sh#L13

The PG_OOM_ADJUST_FILE is set because of OOM killer - in case it hits,
it should kill a child process and not the main one (postmaster).

Setting POSTGRESQL_MAX_CONNECTIONS to 20 has no particular reason,
but the default 100 seems quite high.