Comment out readinessProbe
This is still a bit mystery to me.
After we moved to a new cluster we started seeing something
killing the postgres and the readiness probe was failing.
So the first thing I tried was to remove the probe.
That didn't help so I rulled it out as a cause.
I investigated all memory related stuff for couple weeks!
and in despair tried to again remove the probe and behold,
everything seems to work OK since then.
I don't have energy to investigate it more now because I still
can't believe that after all this has always been the cause.