Update comment about alembic migrations mem usage
I was also thinking about directly increasing the memory limit to 2Gi.
It can be done because we don't have any quota now and we wouldn't need to
increase it manually next (each) time we do an alembic migration.
On the other hand, it moves the packit-service pod from Guaranteed to Burstable QoS,
which means a higher probability of being killed if the node runs out of memory.
Not sure how likely that is, but given that there are AFAICT no quotas
per project, it can IMO happen.
To stay on the safe side, I decided to leave it as it is now.