Give service more memory
After moving service to F35 & mod_wsgi-express we were seeing some errors after the pod was running for some time.
As a workaround we restart the process(es) periodically, which helped on stg, but not on prod.
The culprit seems to be that now we need more memory.
From metrics the top was 247Mi so far, so I'm setting new requests/limits to 320Mi.

https://github.com/packit/packit-service/pull/1363
https://github.com/packit/packit-service/pull/1381