Extract common tasks to separate files
It can't be used everywhere (e.g. together with notify:),
but still, it saves cca 70 lines.

My initial idea was to add something like
block:
  k8s:
    ...
  when: ansible_version.full <= '2.9.27'
  kubernetes.core.k8s:
    ...
  when: ansible_version.full > '2.9.27'

so it'd work on both, F35 & F36.
But it doesn't work in neither of those cases
because you always don't have on of the modules.