Add fluentd sidecar container to worker
https://github.com/packit/research/tree/main/logs-aggregation

We use our image https://quay.io/repository/packit/fluentd-splunk-hec,
built via
https://github.com/jpopelka/fluent-plugin-splunk-hec/blob/main/.github/workflows/rebuild-and-push-image.yml
because we don't want to use docker.io/splunk/fluentd-hec image.

I've been unsuccessful in sending data to Splunk HEC so far.
This is a middle step to make sure the fluentd sidecar works OK
and at this moment we're sending logs to files in a persistent volume
mount to /var/log/packit/ in the fluentd container.
https://docs.fluentd.org/output/file
Once a day is the intermediate buffer.*.log file in /var/log/packit/ flushed
and compressed into a permanent gzip file.
The file is '.' prefixed, so use 'ls -a' to see it.

The fluentd-config has also a match section with '@type splunk_hec',
but it's not matched :) because the preceding '@type file' is matched first.

The fluentd sidecar has no resource requests/limits because with it
the fluentd was starting very slow (about 10sec) and it missed some logs
from worker which was already up and processing.