Don't use tasks/k8s.yml to deploy flower-htpasswd secret
Beside notify (492ccf419a) this is another use case when the tasks/k8s.yml can't be used.

The problem is that the loop item is ALWAYS evaluated even when the 'when:' condition
resolves to False (i.e. in this case when with_flower is False).
https://docs.ansible.com/ansible/latest/user_guide/playbooks_conditionals.html#loops-and-conditionals

When with_flower is False, the htpasswd file is not created in the previous task
and the secret-flower-htpasswd.yml.j2 template then fails to be processed.