Smaller shared_buffers for packit-prod postgres
It a bit worries me that it's been just a couple of months since
I last increased the memory limit from 512Mi to 1Gi (898c2ea5c)
and it looks like it's again not enough.
In the past 2 weeks the memory consumption has been somewhere near the limit,
once it even hit the limit and the process was OOMkilled.

Before we request more memory, let's try to tweak what we have now.
The shared_buffers and effective_cache_size
default values are based on the memory limit [1] and are set reasonably,
(shared_buffers to 25% of the limit and eff_cache_size to 50% of the mem limit)
but there's still space for experiments.
Per [2], the shared_buffers should be set to 15-25% of the limit, and
per [3]: "On systems with less than 1GB of RAM, a smaller percentage of RAM
is appropriate, so as to leave adequate space for the operating system."
So let's try something like 20%, i.e approx. 200Mi.

[1] https://github.com/sclorg/postgresql-container/blob/master/src/root/usr/share/container-scripts/postgresql/common.sh#L12
[2] https://www.enterprisedb.com/postgres-tutorials/how-tune-postgresql-memory
[3] https://www.postgresql.org/docs/current/runtime-config-resource.html#GUC-SHARED-BUFFERS