Add render-secrets-from-templates target
You need it if you want to create packit-service.yaml in order to run
scripts/update_oc_secret.sh packit-config secrets/packit/stg/packit-service.yaml

I've been thinking whether to call it directly after downloading the secrets
from the vault, but leaving it separated so that people are aware that it's
generated and not downloaded from the vault.

Also it's not needed during the deployment so it'd slow it down.