Unify & symplify labels and selectors
Currently we use, 'name', 'app' and 'service' as labels. Per [1]
'name' is recommended for the name of the application
'app' is used in most docs (like in [2]) but not as a name of the
      whole application, but as a name of the individual components
'service' - I haven't seen this in any docs and it's really confusing
            (because of the 'kind: Service').

We're free to use any label we want but from the [1]
the standard label for what we want is probably 'component'.

I've also removed labels where I don't think they're necessary, for example
I don't think we need .metadata.labels for Deployment, Service and Route.

[1] https://www.densify.com/kubernetes-autoscaling/kubernetes-labels
[2] https://kubernetes.io/docs/concepts/workloads/controllers/deployment/