Adjust resource limits based on prod metrics
Looking at packit prod resource utilization graphs,
one can see that we're requesting 1.4cpu,
but utilizing 0.1-0.4cpu, so we're too greedy.

I went through packit-prod metrics for individual components
for past 2 weeks and adjusted the resourse limits to be 2-3x
the values at peaks so we should still be safe.

You can see that it's mostly adjusting cpu, but in some cases
it decreases the memory limit from 80Mi to 32Mi.
Those are components which need only a few MiB of memory
but previously had the limit set to 80Mi because of
Openshift Online where the limit couldn't be less than 80Mi,
see 746cf271692