Yet another update of resource requests & limits
Use Burstable QoS for more/most pods, not just workers (d9278288).

Guaranteed QoS (when requests == limits) is the 'best' class, but you
have to carefully set the values (the same for requests and limits) so that
- it's enough for all use cases, but at the same time
- it's not too much because you'd waste the reserved (requested) resources.

And you actually don't need the generic workloads to be 'quaranteed'.
The Guaranteed QoS makes sense for the 'sensitive' workloads, like in
our case postgres and redis (maybe the service/api).

With the Burstable QoS (requests < limits) we can set
- requests to what the pod needs most of the time
- limits to some high value (more than what the pod needs in spikes)
  to have a safety net when a pod gets crazy
  - when it becomes a memory hog due to a memory leak
  - when it starts eating all cpu because of a bug