Increase liveness probe values
We need to be much less aggressive because the probe
is still failing quite often.

The cause might be the 'solo' pool, which blocks the
worker while it executes tasks.
https://docs.celeryq.dev/en/latest/userguide/workers.html#remote-control

For now, just try a longer timeout (5min should be
enough for all tasks) and if that doesn't help,
we can go back to 'prefork' pool.