Tweak tokman resources and the number of workers
Tokman had issues keeping up with the load on prod, which manifested in
the liveness and readiness probes failing quite often, and occasionally
ridiculous response times.

Using 2 gunicorn workers somewhat improves the situation, but that
requires more resources. Update the playbook and the vars file for
packit-prod accordingly.

Also change from using a readinessProbe to using a startupProbe. For
tokman being ready and being live is the same: a response is received in
a timely manner on /api/health. But we don't want to run the liveness
probe before the startup has completed (the alembic migration completed
and gunicorn started). Use the startup probe to check this. This will
run only once, trying a few times, and stopping once successful. It
might be that this was the intention with the readinessProbe in the
first place :)

Signed-off-by: Hunor Csomortáni <csomh@redhat.com>