chore(postgres): enlarge the postgres volume
Also resized the stage for the testing purposes, it ain't much(, but
it's an honest work), if we reserve a lot of resources on preprod, we
can relax it (we're using barely 300MB on stage).

Turns out that you, indeed, can resize the PVC with the pod running and
then just rescale the deployment (postgres).

Fixes #604

Signed-off-by: Matej Focko <mfocko@redhat.com>